# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MediTime API
  description: |
    API REST complète pour **MediTime**, application de gestion de médicaments et suivi des traitements médicaux.

    ## 🔐 Authentification
    L'API utilise **Supabase Auth** avec tokens JWT Bearer.  
    La plupart des endpoints nécessitent le header `Authorization: Bearer <token>`.

    ### Obtenir un token
    1. Authentifiez-vous via Supabase Auth (email/password, Google, etc.)
    2. Récupérez le token JWT depuis la session Supabase
    3. Ajoutez-le au header : `Authorization: Bearer <votre_token>`

    ## 🏗 Architecture
    | Composant | Technologie |
    |-----------|-------------|
    | Frontend | React 19 (Vite) + Bootstrap 5 + Supabase Client |
    | Backend | Flask (Python) + PostgreSQL + Raw SQL (psycopg2) |
    | Base de données | PostgreSQL avec Row Level Security (RLS) |
    | Infrastructure | Docker + CapRover + Render |

    ## 🚀 Guide de démarrage
    1. `GET /api/status` — Vérifier la connexion
    2. `GET /api/user/sync` — Synchroniser votre profil
    3. `GET /api/calendars` — Lister vos calendriers
    4. `POST /api/calendars` — Créer un calendrier

    ## 📦 Codes de réponse
    | Code | Signification |
    |------|---------------|
    | 200 | Succès |
    | 400 | Requête invalide (données manquantes) |
    | 401 | Non authentifié |
    | 403 | Accès refusé |
    | 404 | Ressource non trouvée |
    | 500 | Erreur serveur |

  version: 2.0.0
  contact:
    name: MediTime Support
...
  license:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT Supabase Auth.
        
        **Format du header :**
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIs...
        ```
  schemas:
    Error:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
          description: Message d'erreur lisible
          example: "Erreur lors de la récupération des données"
        code:
          type: string
          description: Code d'erreur machine
          example: "CALENDAR_FETCH_ERROR"
        i18n_key:
          type: string
          description: Clé de traduction i18n
          example: "api.calendar.fetch_error"
        error:
          type: string
          description: Détails techniques (dev uniquement)
        status_code:
          type: integer
          example: 500
    Success:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
          example: "Opération réussie"
        code:
          type: string
          example: "OPERATION_SUCCESS"
        i18n_key:
          type: string
        data:
          type: object
          additionalProperties: true
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: UID Supabase
        display_name:
          type: string
          example: "Jean Dupont"
        email:
          type: string
          format: email
          example: "jean@example.com"
        photo_url:
          type: string
          format: uri
        email_enabled:
          type: boolean
          description: Notifications email activées
        push_enabled:
          type: boolean
          description: Notifications push activées
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Token:
      type: object
      properties:
        id:
          type: integer
        token:
          type: string
        calendar_id:
          type: integer
        owner_uid:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
...
    Schedule:
      type: object
      properties:
        schedule:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
        table:
          type: object
        calendar_name:
          type: string
        if_low_stock:
          type: boolean
paths:
...
  /user/sync:
    get:
      tags: [User]
      summary: Récupérer le profil utilisateur
      operationId: getUserInfo
      responses:
        '200':
          description: Profil récupéré
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '404':
          description: Utilisateur non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/update:
    put:
      tags: [User]
      summary: Mettre à jour le profil
      operationId: updateUserInfo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                email:
                  type: string
                  format: email
                photo_url:
                  type: string
                  format: uri
                email_enabled:
                  type: boolean
                push_enabled:
                  type: boolean
      responses:
        '200':
          description: Profil mis à jour
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/photo:
    post:
      tags: [User]
      summary: Uploader une photo de profil
      operationId: uploadUserPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo]
              properties:
                photo:
                  type: string
                  format: binary
      responses:
        '200':
          description: Photo uploadée
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          photo_url:
                            type: string
                            format: uri
        '400':
          description: Fichier manquant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ═══════════════════════════════════════════════════════════════════════════
  # CALENDARS
  # ═══════════════════════════════════════════════════════════════════════════
...
  /tokens:
    get:
      tags: [Tokens]
      summary: Lister mes tokens de partage
      operationId: getTokens
      responses:
        '200':
          description: Tokens récupérés
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          tokens:
                            type: array
                            items:
                              $ref: '#/components/schemas/Token'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tokens/{calendar_id}:
    post:
      tags: [Tokens]
      summary: Créer un token de partage
      operationId: createToken
      parameters:
        - name: calendar_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                expiresAt:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        '200':
          description: Token créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Token déjà existant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Accès non autorisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tokens/{token}:
    get:
      tags: [Tokens]
      summary: Métadonnées d'un token public
      operationId: getTokenMetadata
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Métadonnées récupérées
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          calendar_id:
                            type: integer
                          owner_uid:
                            type: string
        '404':
          description: Token non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Tokens]
      summary: Supprimer un token
      operationId: deleteToken
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token supprimé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tokens/expiration/{token}:
    patch:
      tags: [Tokens]
      summary: Modifier l'expiration d'un token
      operationId: updateTokenExpiration
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                expiresAt:
                  type: string
                  format: date
                  nullable: true
      responses:
        '200':
          description: Expiration mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tokens/{token}/schedule:
    get:
      tags: [Tokens]
      summary: Planning via token public
      operationId: getTokenSchedule
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Planning généré
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Schedule'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...